requirements.txt
alembic==1.13.1
async-timeout==4.0.3
Authlib==1.3.1
bidict==0.23.1
blinker==1.8.2
cachecontrol==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
eventlet==0.36.1
firebase-admin==6.5.0
Flask==2.0.2
Flask-Cors==4.0.1
Flask-Dance==3.0.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.5.0
flask-mail==0.10.0
Flask-Migrate==3.1.0
flask-sock==0.7.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
git-filter-repo==2.38.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.64.1
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.2
python-dotenv==0.19.2
python-engineio==4.9.1
python-socketio==5.11.2
redis==5.0.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==1.4.27
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
URLObject==2.4.3
websockets==12.0
Werkzeug==2.0.2
wsproto==1.2.0
wtforms==3.1.2
